Changed cut slider, fix for settings not saving
The transition frame slider now defaults to the middle.

Settings now only save when pressing the ok button in the properties window to avoid a race condition between stinger_update and stinger_path_modified. This could lead to the set transition frame to be ignored and the default setting to be chosen instead.